requirements.txt
Flask
Flask-Session
Werkzeug
mysql-connector-python
python-dotenv
Pillow
requests
